chore: move types to respective files
