+++ src/components/seo.tsx
@@ -3,7 +3,34 @@ import React from "react"
 import { graphql, useStaticQuery } from "gatsby"
 import Helmet from "react-helmet"
 
-import { MetaTag, SEOProps, SiteQueryResult } from "../types"
+interface SiteQueryResult {
+  site: {
+    siteMetadata: {
+      title: string
+      description: string
+      author: string
+      navigationString: string
+      coverImage: string
+      social: Array<{
+        name: string
+        url: string
+      }>
+    }
+  }
+}
+
+interface MetaTag {
+  name?: string
+  property?: string
+  content: string
+}
+
+export interface SEOProps {
+  description?: string
+  lang?: string
+  meta?: MetaTag[]
+  title: string
+}
 
 const SEO: React.FC<SEOProps> = ({
   description = "",
